fix: location.elevation 타입 수정

+++ routee-activity/src/main/java/org/sopt/routee/activity/internal/service/dto/command/CreateTimelineCommand.java
@@ -14,7 +14,7 @@ public record CreateTimelineCommand(
 	Integer trackPointIndex,
 	Double longitude,
 	Double latitude,
-	Double elevation,
+	Integer elevation,
 	Integer pointIndex,
 	TimelineStatus status,
 	ZoneId timeZone
